Update flasky app, continued EDA
