+++ app/templates/profile.html
@@ -4,4 +4,4 @@
 <h1>{{ output1 }} </h1>
 
 
-<input type="button" class="button" value="Melon Button">
+<input type="button" class="button" value="Enter New Infomation">
